@@ -4,13 +4,13 @@ import { ShoppingCartService } from 'app/restaurant-detail/shopping-cart/shoppin
 import { Order } from './order.model';
 import { Observable } from 'rxjs/Observable';
 import 'rxjs/add/operator/map';
-import { Headers, Http, RequestOptions } from '@angular/http';
+import { HttpClient } from '@angular/common/http';
 import { MEAT_API } from 'app/app.api';
 
 @Injectable()
 export class OrderService {
 
-  constructor(private cartService: ShoppingCartService, private http: Http) { }
+  constructor(private cartService: ShoppingCartService, private http: HttpClient) { }
 
   itemsValue():number{
     return this.cartService.totalizacao();
@@ -33,13 +33,8 @@ export class OrderService {
   }
 
   checkOrder(order: Order): Observable<string>{
-    const headers = new Headers();
-    headers.append('Content-Type','application/json')
-    return this.http.post(`${MEAT_API}/orders`, 
-          JSON.stringify(order), 
-          new RequestOptions({headers: headers}))
-          .map(response => response.json())
-          .map(order => order.id);
+    return this.http.post<Order>(`${MEAT_API}/orders`, order)
+    .map(order => order.id);
   }
 
   clear() {
